requirements.txt
# Core dependencies
pydantic-ai==0.0.14
pydantic==2.9.2

# Logfire for observability
logfire==0.58.0

# HTTP requests for web scraping
httpx==0.27.2
requests==2.32.3

# Environment variables
python-dotenv==1.0.1

# Additional utilities
beautifulsoup4==4.12.3
lxml==5.3.0

# Google Gemini API
google-generativeai==0.3.2

# MCP (Model Context Protocol) client
mcp

# Context7 for real web search via MCP
# Install with: pip install git+https://github.com/context7/mcp-server-context7.git
